final.final? review cleanup
